修改

+++ icmes-information-service/src/main/java/com/dtelec/icmes/information/service/DeviceCategoryCommandHandler.java
@@ -1,13 +1,17 @@
 package com.dtelec.icmes.information.service;
 
+import java.util.List;
 import java.util.UUID;
 
 import org.apache.commons.lang.StringUtils;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DuplicateKeyException;
 import org.springframework.stereotype.Service;
 import com.dtelec.icmes.common.error.IcmesBusinessException;
 import com.dtelec.icmes.common.error.IcmesErrorTypeEnum;
 import com.dtelec.icmes.information.repository.IDeviceCategoryRepository;
+import com.dtelec.icmes.information.repository.IDeviceRepository;
+import com.dtelec.icmes.information.repository.entity.DeviceBaseEntity;
 import com.dtelec.icmes.information.repository.entity.DeviceCategoryBaseEntity;
 import com.dtelec.icmes.information.service.annotation.CommandAction;
 import com.dtelec.icmes.information.service.command.DeviceCategoryCreateCommand;
@@ -27,6 +31,8 @@ public class DeviceCategoryCommandHandler implements ICommandHandler {
 	@Autowired
 	private IDeviceCategoryRepository deviceCategoryRepo;
 	
+	@Autowired
+	private IDeviceRepository deviceRepo;
 	/**
 	 * 设备类型新建
 	 * @param command
@@ -42,8 +48,8 @@ public class DeviceCategoryCommandHandler implements ICommandHandler {
 	    		
 				deviceCategoryRepo.createDeviceCategory(entity);
 				
-	    	}catch(Exception e){
-	    		throw new IcmesBusinessException(IcmesErrorTypeEnum.INFO_DEVICECATEGORY_CREATEDEVICEVATEGORY_ISNOTNULL, "设备编号已存在");
+	    	}catch(DuplicateKeyException e){
+	    		throw new IcmesBusinessException(IcmesErrorTypeEnum.INFO_DEVICECATEGORY_CREATEDEVICEVATEGORY_ISNOTNULL, "设备类型编号已存在");
 	    	}
 			
 			deviceCategoryRepo.addSpecDataForDeviceCategory(entity.getId(), command.getRelatedSpceDataIds());
@@ -101,7 +107,10 @@ public class DeviceCategoryCommandHandler implements ICommandHandler {
 			throw new IcmesBusinessException(IcmesErrorTypeEnum.INFO_DEVICECATEGORY_CREATEDEVICEVATEGORY_ISNULL, "该设备类型不存在或已删除");
 		}
 		
-//		List<>findDevicesByCategoryId();
+		List<DeviceBaseEntity> devices = deviceRepo.findDevicesByDeviceCategoryId(id);
+		if (devices != null && !devices.isEmpty()) {
+			throw new IcmesBusinessException(IcmesErrorTypeEnum.INFO_DEVICECATEGORY_DELETEDEVICEVATEGORY_ISFAIL, "设备类型与设备有关联，不能删除");
+		}
 		
 		deviceCategoryRepo.deleteDeviceCategory(id);
